requirements.txt
Django>=4.2
djangorestframework
pandas
requests
psycopg2-binary
